Fix duplicated test

+++ chainerx_cc/chainerx/routines/type_util_test.cc
@@ -173,7 +173,7 @@ TEST(ResultTypeTest, Three) {
     CHECK_RESULT_TYPE_HOMO3(Int32, i8, i8, i32);
     CHECK_RESULT_TYPE_HOMO3(Int32, i8, i16, i32);
     CHECK_RESULT_TYPE_HOMO3(Int32, i8, i32, i32);
-    CHECK_RESULT_TYPE_HOMO3(Int32, i8, i32, i32);
+    CHECK_RESULT_TYPE_HOMO3(Int64, i8, i64, i32);
 
     // unsigned ints
     CHECK_RESULT_TYPE_HOMO3(UInt8, u8, u8, u8);
